// frontend/smartiphr/src/app/models/registroCarrello.ts
export class RegistroCarrello {
  _id?: String;
  operator?: String;
  type?: String;
  operatorName?: String;
  paziente?: String;
  pazienteName?: String;
  dataModifica?: Date;
  operation?: String;
  elemento?: String;
  carrelloID?: String;
  carrelloName?: String;
  quantita?: Number;
  quantitaRes?: Number;
}